djangoProject/requirements.txt
django-cors-headers
django_rest_framework